blog/requirements.txt
django
psycopg2-binary
django-widget-tweaks
django-oscar
django-oscar-api
django-compressor
flake8
ipython
django-nose
